+++ src/lib/utils/zustandPersistNames.ts
@@ -1,0 +1,3 @@
+export const zNames = {
+  paletteFlag: "z-devtoggle-palette",
+}
