+++ src/palette/PaletteFlag.tsx
@@ -1,25 +1,8 @@
-import AsyncStorage from "@react-native-community/async-storage"
 import { MenuItem } from "lib/Components/MenuItem"
 import { Text } from "palette"
 import React from "react"
 import { Alert } from "react-native"
-import create from "zustand"
-import { persist } from "zustand/middleware"
-
-interface PaletteFlagState {
-  allowV3: boolean
-  setAllowV3: (value: boolean) => void
-}
-
-export const usePaletteFlagStore = create<PaletteFlagState>(
-  persist(
-    (set) => ({
-      allowV3: __TEST__,
-      setAllowV3: (value) => set((_state) => ({ allowV3: value })),
-    }),
-    { name: "z-devtoggle-palette", getStorage: () => AsyncStorage }
-  )
-)
+import { usePaletteFlagStore } from "./PaletteFlagStore"
 
 export const DevTogglePaletteFlag = () => {
   const currentValue = usePaletteFlagStore((state) => state.allowV3)
